refactor: delegate files to its correct card folder
